+++ src/server/services/dbo/index.ts
@@ -1,3 +1,2 @@
-export * from './mongoPersistance';
 export * from './postgreSqlPersist';
 export * from './iPersist';